refactor: mediaViewer 컴포넌트 내부 로직 분리
