+++ apps/web/app/review-modal/components/ImageCarousel.tsx
@@ -1,0 +1,68 @@
+'use client';
+
+import type { Swiper as SwiperType } from 'swiper';
+import 'swiper/css';
+import { Swiper, SwiperSlide } from 'swiper/react';
+import { useRef, useState } from 'react';
+import Image from 'next/image';
+import { IconButton } from '@lococo/design-system';
+import { SvgArrowUp, SvgArrowDown } from '@/icons';
+
+interface ImageCarouselProps {
+  images: string[];
+}
+
+export default function ImageCarousel({ images }: ImageCarouselProps) {
+  const swiperRef = useRef<SwiperType | null>(null);
+  const [currentIndex, setCurrentIndex] = useState(0);
+
+  const navBtnClass =
+    'border-1 size-[3.2rem] -rotate-90 border-gray-200 bg-white';
+
+  return (
+    <>
+      <Swiper
+        direction="horizontal"
+        onSwiper={(s) => (swiperRef.current = s)}
+        onSlideChange={(s) => setCurrentIndex(s.activeIndex)}
+        className="absolute inset-0 h-full w-full"
+      >
+        {images.map((img, i) => (
+          <SwiperSlide key={i}>
+            <Image
+              src={img}
+              alt="리뷰 이미지"
+              fill
+              className="relative h-full w-full object-cover"
+            />
+          </SwiperSlide>
+        ))}
+      </Swiper>
+
+      {currentIndex > 0 && (
+        <div className="absolute left-[1.2rem] top-1/2 z-20 -translate-y-1/2">
+          <IconButton
+            icon={<SvgArrowUp />}
+            rounded
+            color="secondary"
+            aria-label="이전 이미지"
+            onClick={() => swiperRef.current?.slidePrev()}
+            className={navBtnClass}
+          />
+        </div>
+      )}
+      {currentIndex < images.length - 1 && (
+        <div className="absolute right-[1.2rem] top-1/2 z-20 -translate-y-1/2">
+          <IconButton
+            icon={<SvgArrowDown />}
+            rounded
+            color="secondary"
+            aria-label="다음 이미지"
+            onClick={() => swiperRef.current?.slideNext()}
+            className={navBtnClass}
+          />
+        </div>
+      )}
+    </>
+  );
+}
