+++ apps/web/app/review-modal/components/ReviewModalLayout.tsx
@@ -26,7 +26,7 @@ export default function ReviewModalLayout({
   children,
 }: ReviewModalLayoutProps) {
   return (
-    <div className="fixed inset-0 z-20 flex items-center justify-center">
+    <div className="relative flex h-full w-full items-center justify-center">
       <div className="relative flex h-[55.2rem] w-[97.2rem] rounded-xl bg-white">
         {/* 좌측: 미디어 + 하단정보 */}
         <div className="relative flex w-[55.2rem] flex-col">
